requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
g4f
psutil>=5.9.0
asyncpg>=0.28.0
yt-dlp>=2023.3.4
PyNaCl>=1.5.0
wavelink>=2.0.0
aiohttp>=3.8.0
async-timeout
psycopg2-binary
ffmpeg-python
mutagen
pydub
